Function

Writes the files for a remake workflow. Use the help_remakeGenerator function to get more help.

workflow(targets = NULL, sources = NULL, packages = NULL,
  remakefile = "remake.yml", makefile = "Makefile", begin = NULL,
  clean = NULL, remake_args = list())

Arguments

targets

YAML-like list of targets, which you can generate by supplying data frames of remake commands to the targets function.

sources

Character vector of R source files

packages

Character vector of packages

remakefile

Character, name of the remake file to generate. Should be in the current working directory.

makefile

Character, name of the Makefile. Should be in the current working directory. Set to NULL to suppress the writing of the Makefile.

begin

Character vector of lines to prepend to the Makefile.

clean

Character vector of extra Makefile commands for make clean.

remake_args

Fully-named list of additional arguments to remake::make. You cannot set target_names or remake_file this way because those names are reserved.

Details

Use the help_remakeGenerator function to get more help.

See also

help_remakeGenerator